# api-gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorDetail:
    properties:
      message:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorDetail'
    type: object
...
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.IDResponse:
    properties:
      id:
        type: string
    type: object
...
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.PricingRule:
    properties:
      class:
        type: string
      createdAt:
        type: string
      freeMinutes:
        type: integer
      id:
        type: string
      isActive:
        type: boolean
      minChargeTenge:
        type: integer
      modelID:
        type: string
      overtimePolicy:
        type: string
      overtimeRateTenge:
        type: integer
      ratePerKMTenge:
        type: integer
      rateTenge:
        type: integer
      type:
        type: string
      updatedAt:
        type: string
      zoneID:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.PricingRuleCreateRequest:
    properties:
      class:
        type: string
      freeMinutes:
        type: integer
      minChargeTenge:
        type: integer
      modelID:
        type: string
      overtimePolicy:
        type: string
      overtimeRateTenge:
        type: integer
      ratePerKMTenge:
        type: integer
      rateTenge:
        type: integer
      type:
        type: string
      zoneID:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.PricingRuleResponse:
    properties:
      rule:
        $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.PricingRule'
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.PricingRuleUpdateRequest:
    properties:
      class:
        type: string
      freeMinutes:
        type: integer
      isActive:
        type: boolean
      minChargeTenge:
        type: integer
      modelID:
        type: string
      overtimePolicy:
        type: string
      overtimeRateTenge:
        type: integer
      ratePerKMTenge:
        type: integer
      rateTenge:
        type: integer
      type:
        type: string
      zoneID:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.PricingRulesResponse:
    properties:
      rules:
        items:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.PricingRule'
        type: array
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@carrental.kz
    name: API Support
  description: API Gateway for the Kazakhstan carsharing platform. Handles auth, users,
    fleet, zones, insurance, and maintenance.
  license:
    name: Proprietary
  termsOfService: http://swagger.io/terms/
  title: Car Rental API Gateway
  version: "1.0"
paths:
...
  /pricing-rules:
    get:
      description: Returns pricing rules filtered by optional scope dimensions.
      parameters:
      - description: Filter by car model UUID
        in: query
        name: modelID
        type: string
      - description: Filter by zone UUID
        in: query
        name: zoneID
        type: string
      - description: Filter by car class
        in: query
        name: class
        type: string
      - description: Filter by pricing type
        in: query
        name: type
        type: string
      - description: Filter by active status
        in: query
        name: isActive
        type: boolean
      - description: Pagination limit
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.PricingRulesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List pricing rules
      tags:
      - pricing-rules
    post:
      consumes:
      - application/json
      description: Defines a new pricing rule. Nil scope fields (modelID, zoneID,
        class) act as wildcards and apply to all values of that dimension.
      parameters:
      - description: Pricing rule create payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.PricingRuleCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a pricing rule
      tags:
      - pricing-rules
  /pricing-rules/{id}:
    delete:
      parameters:
      - description: Pricing rule UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete pricing rule
      tags:
      - pricing-rules
    get:
      parameters:
      - description: Pricing rule UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.PricingRuleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get pricing rule by ID
      tags:
      - pricing-rules
    patch:
      consumes:
      - application/json
      description: Partially updates a pricing rule. Only provided fields are changed.
      parameters:
      - description: Pricing rule UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.PricingRuleUpdateRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update pricing rule
      tags:
      - pricing-rules
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"